# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Ficheiro de Clientes - Restaurant Loyalty System
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItemType } from './activityItemType';

export interface ActivityItem {
  type: ActivityItemType;
  customerName: string;
  description: string;
  timestamp: string;
  /** @nullable */
  tier?: string | null;
}